+++ .env.sample
@@ -17,3 +17,6 @@ EMAIL_HOST=pochtoviyserversov@yandex.ru
 EMAIL_PORT=
 EMAIL_HOST_USER=
 EMAIL_HOST_PASSWORD=
+
+REDIS_HOST=
+REDIS_PORT=
