+++ lms/tasks.py
@@ -1,14 +1,8 @@
-from datetime import timedelta
-
 from celery import shared_task
 from django.conf import settings
-from django.contrib.auth import get_user_model
 from django.core.mail import send_mail
-from django.utils import timezone
 
 from lms.models import Course, Subscription
-
-User = get_user_model()
 
 
 @shared_task
@@ -47,20 +41,3 @@ def send_course_update_notification(course_id):
         return f"Course with id {course_id} does not exist"
     except Exception as e:
         return f"Error sending notifications: {str(e)}"
-
-
-@shared_task
-def block_inactive_users():
-    """
-    Задача для блокировки пользователей, которые не заходили более месяца
-    """
-    try:
-        month_ago = timezone.now() - timedelta(days=30)
-        inactive_users = User.objects.filter(last_login__lt=month_ago, is_active=True)
-
-        count = inactive_users.count()
-        inactive_users.update(is_active=False)
-        return f"Заблокировано {count} неактивных пользователей"
-
-    except Exception as e:
-        return f"Ошибка при блокировке пользователей: {str(e)}"
