+++ users/tasks.py
@@ -1,0 +1,20 @@
+from datetime import timedelta
+from celery import shared_task
+from django.contrib.auth import get_user_model
+from django.utils import timezone
+
+
+User = get_user_model()
+
+
+@shared_task
+def block_inactive_users():
+    """
+    Задача для блокировки пользователей, которые не заходили более месяца
+    """
+
+    month_ago = timezone.now() - timedelta(days=30)
+    inactive_users = User.objects.filter(last_login__lt=month_ago, is_active=True)
+
+    count = inactive_users.update(is_active=False)
+    return f"Заблокировано {count} неактивных пользователей"
